requirements.txt
python-telegram-bot>=21.0
Flask>=3.0.0
gunicorn>=21.0
requests>=2.31.0
sseclient-py>=1.8.0
